net: ipc_router: Fix memory leaks when releasing a remote port
In ipc_router_set_conn(), memory is allocated to a local port for the
conn_info linked list, but the memory for the local ports is never
freed when the remote port is released.

Delete the local ports and free all of their memory when releasing a remote
port to fix the memory leaks. This is just like how memory is freed for the
remote_tx ports.

Change-Id: I136142abb3dc877db441b3fcc1f368b3d36b9dea
Signed-off-by: Sultanxda <sultanxda@gmail.com>
(cherry picked from commit 0e25207d52c684ab05eb5ff84c5136e420ee31db)